Added validation (authentication and edit before expirationDate) to ordersPack edition.

+++ server/src/routes.controllers/ordersPack.routes.controller.ts
@@ -6,6 +6,7 @@ import { OrdersPackBuilder } from "../controllers/OrdersPack/OrdersPack.builder"
 import { OrdersPackRepository } from "../controllers/OrdersPack/OrdersPack.repository";
 import { Order } from "../controllers/Order/Order";
 import { OrderRepository } from "../controllers/Order/Order.repository";
+import { Ref } from "@typegoose/typegoose";
 
 export class OrdersPackRoutesController{
 
@@ -21,17 +22,32 @@ export class OrdersPackRoutesController{
     }
 
     public static async updateOrdersPack(req: Request, res: Response): Promise<void>{
-        const {newExpirationDate, deleteArray, ordersPack_id} = req.body;
-        
-        if(newExpirationDate){
-            const newDate: Date =  new Date(Number(newExpirationDate));
-            OrdersPackRepository.addTime(newDate, ordersPack_id);
+        const {newExpirationDate, deleteArray, ordersPack_id, user_id} = req.body;
+        const ordersPack: OrdersPack =  await OrdersPackRepository.findOne(ordersPack_id);
+
+        if(OrdersPackRoutesController.canUpdateOrdersPack(ordersPack, user_id)){
+            if(newExpirationDate){
+                const newDate: Date =  new Date(Number(newExpirationDate));
+                OrdersPackRepository.addTime(newDate, ordersPack_id);
+            }
+    
+            if(deleteArray){
+                OrdersPackRepository.deleteOrders(deleteArray, ordersPack_id);
+            }
+            res.sendStatus(200);
+            return;
         }
 
-        if(deleteArray){
-            OrdersPackRepository.deleteOrders(deleteArray, ordersPack_id);
-        }
-        res.sendStatus(200);
+        res.sendStatus(400);
+        
+    }
+
+    private static canUpdateOrdersPack(ordersPack: OrdersPack, user_id: Ref<User>): boolean{
+        return(
+            ordersPack &&
+            ordersPack.creator['_id'] == user_id &&
+            ordersPack.expirationDate > new Date()
+        );
     }
 
     public static async deleteOrdersPack(req: Request, res: Response): Promise<void>{
